requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.4
arch==7.0.0
asttokens==2.4.1
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
Backtesting==0.3.3
beautifulsoup4==4.12.3
bibtexparser==2.0.0b5
blinker==1.6.2
bokeh==3.4.1
bottle==0.12.25
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.3
clldutils==3.22.1
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.0
comm==0.2.2
conan==2.4.1
contourpy==1.2.1
cryptography==40.0.2
csvw==3.2.2
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dlinfo==1.2.1
dnspython==2.3.0
docstring_parser==0.16
executing==2.0.1
extras==1.0.0
fasteners==0.19
filelock==3.13.1
fire==0.6.0
Flask==2.3.2
fonttools==4.52.4
frozendict==2.4.4
fsspec==2023.12.2
future==0.18.3
google-auth==2.26.1
google-auth-oauthlib==1.2.0
gpuctypes==0.2.0
grpcio==1.60.0
gym==0.26.2
gym-notices==0.0.8
h11==0.14.0
hmmlearn==0.3.2
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface==0.0.1
huggingface-hub==0.20.2
idna==3.6
impacket==0.10.0
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
language-tags==1.2.0
lazy_loader==0.3
ldap3==2.9.1
ldapdomaindump==0.9.4
librosa==0.10.1
lightweight-charts==1.0.22
llvmlite==0.41.1
lxml==5.1.0
Markdown==3.5.1
MarkupSafe==2.1.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mistral_common==1.2.1
mistral_inference==1.1.0
mpmath==1.3.0
msgpack==1.0.7
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
openai==1.6.1
packaging==23.2
pandas==2.2.2
parso==0.8.4
patch-ng==1.17.4
patsy==0.5.6
peewee==3.17.5
pexpect==4.9.0
phonemizer==3.2.1
pillow==10.3.0
platformdirs==4.1.0
plotly==5.22.0
plotly-express==0.4.1
pooch==1.8.0
prompt_toolkit==3.0.45
protobuf==4.23.4
proxy_tools==0.1.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodomex==3.18.0
pycurl==7.45.2
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.18.0
pylatexenc==2.10
pyobjc-core==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-libdispatch==10.1
pyobjc-framework-Metal==10.1
pyobjc-framework-Security==10.3
pyobjc-framework-WebKit==10.3
pyOpenSSL==23.1.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
pywebview==4.4.1
PyYAML==6.0.1
pyzmq==26.0.3
rdflib==7.0.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.16.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
segments==2.2.1
sentencepiece==0.1.99
simple_parsing==0.1.5
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
stack-data==0.6.3
statsmodels==0.14.2
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
-e git+https://github.com/tinygrad/tinygrad.git@54629b56d2533ff1c3f604e163746573b13acb36#egg=tinygrad
tokenizers==0.15.0
torch==2.3.1
torch-lr-finder==0.2.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.39.0
typing_extensions==4.12.2
tzdata==2024.1
unicode==2.9
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==1.26.19
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.3.4
xformers==0.0.26.post1
xyzservices==2024.4.0
yfinance==0.2.40
